phase2/Perplexity/requirements.txt
numpy==2.1.1
scipy==1.14.1
Pillow==10.4.0
pytest==8.3.2